requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.5
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
tortoise-orm==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
